remove run_worker_loop, impl poll for WebWorker
